Rename IPC files
